It looks like texting bubbles now
Entirely reworked the stylesheet to look like a phone text message interface with grey bubbles on the left side for messages received and blue bubbles on the right side for messages sent. Also added support for MMS text and images, but I only tested jpegs. It should support multiple contacts (one whole conversation after the other), but I only tested with a single conversation export. The widths are based on my need to print many pages of text messages on ANSI Letter paper.